Deleted UserPage and added LocationPage table
